# docs/04-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: LabelX API
  version: 0.1.0
  description: |
    Quality Control cho annotation CVAT — contract M13 v1.

    Xác thực bằng session LabelX (cookie `sessionid`); request ghi gửi header `X-CSRFToken`
    lấy từ cookie `csrftoken` (gọi `GET /api/auth/csrf/` trước khi đăng nhập).
    Mọi lỗi trả `Error` với `code`, `message`, `request_id` (SRS §9.3.1).
    Đường dẫn có dấu `/` cuối theo quy ước Django REST framework; `/api/snapshots/` trong
    contract này là `/api/snapshots` của SRS tab:api.
servers:
  - url: http://localhost:8000
    description: Backend dev
...
paths:
...
  /api/auth/login/:
    post:
      operationId: auth_login
      tags: [auth]
      summary: Đăng nhập, tạo session
      security: []
      x-labelx-trace: [FR-SEC-01, FR-SEC-05, UC-13]
      x-labelx-roles: [anonymous]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Đăng nhập thành công; trả session hiện tại
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/auth/logout/:
    post:
      operationId: auth_logout
      tags: [auth]
      summary: Đăng xuất, huỷ session
      description: Lease đang giữ không bị trả; nó hết hạn theo FR-REV-04.
      x-labelx-trace: [FR-SEC-01, FR-SEC-05]
      x-labelx-roles: [authenticated]
      responses:
        "204":
          description: Đã đăng xuất (gọi lại khi đã đăng xuất cũng trả 204)
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/auth/session/:
    get:
      operationId: auth_session
      tags: [auth]
      summary: Người dùng, vai trò theo scope, identity mapping
      description: |
        Frontend dùng để guard route và ẩn menu theo vai trò. Ẩn menu chỉ là tiện ích;
        backend vẫn kiểm quyền ở mọi endpoint (FR-SEC-01).
      x-labelx-trace: [FR-SEC-01, FR-SEC-02]
      x-labelx-roles: [authenticated]
      responses:
        "200":
          description: Session hiện tại
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/auth/workflow-permissions/:
    get:
      operationId: auth_workflow_permissions
      tags: [auth]
      summary: Ma trận phân quyền và quy tắc workflow
      description: |
        Cung cấp cấu hình ma trận phân quyền theo vai trò (7 vai trò) và các quy tắc workflow
        (anti-self-review, super admin override, separation of duties) phục vụ màn hình WorkflowPermissions.
      x-labelx-trace: [FR-SEC-01, FR-SEC-06]
      x-labelx-roles: [qa_lead, qc_admin, super_admin]
      responses:
        "200":
          description: Ma trận phân quyền và quy tắc workflow
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowPermissions" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # -------------------------------------------------------------- datasets
...
  /api/guidelines/rules/:
    get:
      operationId: guidelines_rules_list
      tags: [guidelines]
      summary: Tra rule theo nhóm lỗi, lớp, cặp lớp
      description: Không dùng retrieval ngữ nghĩa (FR-GDL-04). Mặc định guideline version mới nhất.
      x-labelx-trace: [FR-GDL-04, UC-12]
      x-labelx-roles: [reviewer, qa_lead, qc_admin, super_admin]
      parameters:
        - $ref: "#/components/parameters/GuidelineVersion"
        - name: family
          in: query
          schema: { $ref: "#/components/schemas/IssueFamily" }
        - name: class_name
          in: query
          schema: { type: string }
        - name: paired_class
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Trang rule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedGuidelineRuleList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/guidelines/rules/{rule_id}/:
    get:
      operationId: guidelines_rules_retrieve
      tags: [guidelines]
      summary: Rule theo guideline version của snapshot
      x-labelx-trace: [FR-GDL-03, FR-GDL-04, FR-REV-07, UC-12]
      x-labelx-roles: [reviewer, qa_lead, qc_admin, super_admin]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema: { type: string, example: VEH-03 }
        - $ref: "#/components/parameters/GuidelineVersion"
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuidelineRule" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- review
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sessionid
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: Con trỏ trang từ trường next/previous (CursorPagination, PAGE_SIZE=50)
      schema: { type: string }
...
    GuidelineVersion:
      name: version
      in: query
      description: Guideline version tag; bỏ trống là bản mới nhất. Workspace truyền version của snapshot.
      schema: { type: string }
  responses:
    BadRequest:
      description: Dữ liệu không hợp lệ (VALIDATION_ERROR, INVALID_CREDENTIALS)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        Chưa đăng nhập (NOT_AUTHENTICATED), không có quyền (FORBIDDEN), ngoài scope (OUT_OF_SCOPE),
        self-review (SELF_REVIEW_FORBIDDEN), tách nhiệm vụ (SAME_REQUESTER_APPROVER),
        thiếu identity mapping (IDENTITY_MAPPING_MISSING). Được ghi audit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Không tìm thấy (NOT_FOUND)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - INVALID_CREDENTIALS
        - NOT_AUTHENTICATED
        - FORBIDDEN
        - OUT_OF_SCOPE
        - SELF_REVIEW_FORBIDDEN
        - SAME_REQUESTER_APPROVER
        - IDENTITY_MAPPING_MISSING
        - NOT_FOUND
        - LEASE_CONFLICT
        - INVALID_TRANSITION
        - REVISION_UNCHANGED
        - IDEMPOTENCY_KEY_REUSED
        - SCOPE_BUSY
        - ISSUE_EXISTS
        - BUSINESS_RULE_UNMET
        - INSUFFICIENT_SAMPLE
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message:
          type: string
          description: Thông báo tiếng Việt cho người dùng
        request_id:
          type: string
          description: Trùng header X-Request-ID của phản hồi; dùng để tra log và audit
        details:
          type: object
          additionalProperties: true
          description: |
            VALIDATION_ERROR: {"fields": {"<tên trường>": ["<lỗi>"]}}.
            INVALID_TRANSITION: {"from": "...", "event": "..."} và có thể drift_jobs.
            BUSINESS_RULE_UNMET ở frames_complete: {"undecided_issue_ids": [..]}.
      example:
        code: LEASE_CONFLICT
        message: Lease đã hết hạn hoặc thuộc người khác
        request_id: 0b6f7c1e-3c55-4d43-9a43-3f0e1f6d2a10
        details: {}

    # ------------------------------------------------------------------ auth
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password, writeOnly: true }
    Role:
      type: string
      description: Vai trò theo SRS tab:rbac và đoạn dưới bảng (Product Owner, Data/Model Owner)
      enum: [annotator, reviewer, qa_lead, qc_admin, super_admin, product_owner, data_model_owner]
    RoleAssignment:
      type: object
      required: [role, dataset_id]
      properties:
        role: { $ref: "#/components/schemas/Role" }
        dataset_id:
          type: integer
          nullable: true
          description: null nghĩa là toàn hệ thống (chỉ super_admin, qc_admin)
    Session:
      type: object
      required: [user, roles, identity_mapping]
      properties:
        user:
          type: object
          required: [id, username, display_name]
          properties:
            id: { type: integer }
            username: { type: string }
            display_name: { type: string }
        roles:
          type: array
          items: { $ref: "#/components/schemas/RoleAssignment" }
        identity_mapping:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [mapped, missing]
              description: missing → mọi thao tác review/phân xử/verify bị từ chối (fail closed)
            cvat_user_id: { type: integer, nullable: true }
    WorkflowPermissions:
      type: object
      required: [matrix, rules, user_roles]
      properties:
        matrix:
          type: array
          items: { $ref: "#/components/schemas/WorkflowMatrixItem" }
        rules:
          type: array
          items: { $ref: "#/components/schemas/WorkflowRule" }
        user_roles:
          type: array
          items: { $ref: "#/components/schemas/RoleAssignment" }
    WorkflowMatrixItem:
      type: object
      required: [function, roles]
      properties:
        function: { type: string }
        roles:
          type: object
          required: [annotator, reviewer, qa_lead, qc_admin, super_admin, product_owner, data_model_owner]
          properties:
            annotator: { type: string }
            reviewer: { type: string }
            qa_lead: { type: string }
            qc_admin: { type: string }
            super_admin: { type: string }
            product_owner: { type: string }
            data_model_owner: { type: string }
    WorkflowRule:
      type: object
      required: [rule, name, description, enforced]
      properties:
        rule: { type: string }
        name: { type: string }
        description: { type: string }
        enforced: { type: boolean }

    # -------------------------------------------------------------- datasets
...
    GuidelineRule:
      type: object
      required: [rule_id, section, content, guideline_version]
      properties:
        rule_id: { type: string, example: VEH-03 }
        section: { type: string }
        content: { type: string }
        guideline_version: { type: string }

    # ---------------------------------------------------------------- review
...
    IssueFamily:
      type: string
      description: E1 thiếu box, E2 sai lớp, E3 trùng box; structural là cảnh báo cấu trúc có rule ID
      enum: [E1, E2, E3, structural]
...
    PaginatedGuidelineRuleList:
      type: object
      required: [results]
      properties:
        next: { type: string, nullable: true }
        previous: { type: string, nullable: true }
        results:
          type: array
          items: { $ref: "#/components/schemas/GuidelineRule" }
security:
  - sessionAuth: []